refactor: split progression logic & fix progression calculations (#126)
