+++ src/game-logic/buildings/carnationShop.ts
@@ -1,5 +1,5 @@
 import CarnationShop from '@/components/models/buildings/CarnationShop.vue'
-import { LinearLevelProgression } from '@/game-logic/level-progression.js'
+import { LinearLevelProgression } from '@/game-logic/level-progression/linear-progression.js'
 import { ResourceRecord, resourcesPerHour } from '@/game-logic/resources.js'
 import { Duration } from '@/utils/duration.js'
 import Big from 'big.js'
