+++ src/game-logic/types.ts
@@ -1,4 +1,4 @@
-import type { LevelProgression } from '@/game-logic/level-progression.js'
+import type { LevelProgression } from '@/game-logic/level-progression/base.js'
 import type { ResourceRecord } from '@/game-logic/resources.js'
 import type { Duration } from '@/utils/duration.js'
 import type { Component } from 'vue'
